+++ doc/python/conf.py
@@ -15,6 +15,7 @@ import corrade.utility
 
 import magnum
 import magnum.gl
+import magnum.materialtools
 import magnum.meshtools
 import magnum.platform
 import magnum.platform.egl
@@ -31,7 +32,7 @@ import magnum.trade
 # So the doc see everything
 # TODO: use just +=, m.css should reorder this on its own
 corrade.__all__ = ['containers', 'pluginmanager', 'utility', 'BUILD_DEPRECATED', 'BUILD_STATIC', 'BUILD_MULTITHREADED', 'TARGET_UNIX', 'TARGET_APPLE', 'TARGET_IOS', 'TARGET_IOS_SIMULATOR', 'TARGET_WINDOWS', 'TARGET_WINDOWS_RT', 'TARGET_EMSCRIPTEN', 'TARGET_ANDROID']
-magnum.__all__ = ['math', 'gl', 'meshtools', 'platform', 'primitives', 'shaders', 'scenegraph', 'scenetools', 'text', 'trade', 'BUILD_DEPRECATED', 'BUILD_STATIC', 'TARGET_GL', 'TARGET_GLES', 'TARGET_GLES2', 'TARGET_WEBGL', 'TARGET_EGL', 'TARGET_VK'] + magnum.__all__
+magnum.__all__ = ['math', 'gl', 'materialtools', 'meshtools', 'platform', 'primitives', 'shaders', 'scenegraph', 'scenetools', 'text', 'trade', 'BUILD_DEPRECATED', 'BUILD_STATIC', 'TARGET_GL', 'TARGET_GLES', 'TARGET_GLES2', 'TARGET_WEBGL', 'TARGET_EGL', 'TARGET_VK'] + magnum.__all__
 
 # hide values of the preprocessor defines to avoid confusion by assigning a
 # class without __repr__ to them
@@ -214,6 +215,7 @@ INPUT_DOCS = [
     'magnum.rst',
     'magnum.gl.rst',
     'magnum.math.rst',
+    'magnum.materialtools.rst',
     'magnum.meshtools.rst',
     'magnum.platform.rst',
     'magnum.scenegraph.rst',